+++ packages/analytics-js-integrations/src/integrations/Optimizely/browser.js
@@ -2,6 +2,8 @@
 import { NAME, DISPLAY_NAME } from './constants';
 import Logger from '../../utils/logger';
 import { mapRudderPropsToOptimizelyProps } from './utils';
+import { isArray } from '../../utils/utils';
+import { isDefinedAndNotNullAndNotEmpty } from '../../utils/commonUtils';
 
 const logger = new Logger(DISPLAY_NAME);
 
@@ -51,23 +53,32 @@ class Optimizely {
     return undefined;
   };
 
+  getAudienceIdsAndNames = audiences => {
+    const audienceIds = [];
+    const audienceNames = [];
+    if (!isArray(audiences)) {
+      return { audienceIds, audienceNames };
+    }
+
+    for (const audience of audiences) {
+      if (isDefinedAndNotNullAndNotEmpty(audience?.id)) {
+        audienceIds.push(audience.id);
+      }
+      if (isDefinedAndNotNullAndNotEmpty(audience?.name)) {
+        audienceNames.push(audience.name);
+      }
+    }
+    return {
+      audienceIds: audienceIds.sort((a, b) => a.localeCompare(b)),
+      audienceNames: audienceNames.sort((a, b) => a.localeCompare(b)),
+    };
+  };
+
   sendDataToRudder = campaignState => {
     const { experiment, variation } = campaignState;
     const context = { integrations: { All: true } };
     const { audiences, campaignName, id, isInCampaignHoldback } = campaignState;
-
-    // Reformatting this data structure into hash map so concatenating variation ids and names is easier later
-    const audiencesMap = {};
-    audiences.forEach(audience => {
-      audiencesMap[audience.id] = audience.name;
-    });
-
-    const audienceIds = Object.keys(audiencesMap)
-      .sort((a, b) => a.localeCompare(b))
-      .join();
-    const audienceNames = Object.values(audiencesMap)
-      .sort((a, b) => a.localeCompare(b))
-      .join(', ');
+    const { audienceIds, audienceNames } = this.getAudienceIdsAndNames(audiences);
 
     if (this.sendExperimentTrack) {
       let props = {
@@ -77,8 +88,8 @@ class Optimizely {
         experimentName: experiment.name,
         variationName: variation.name,
         variationId: variation.id,
-        audienceId: audienceIds, // eg. '7527562222,7527111138'
-        audienceName: audienceNames, // eg. 'Peaky Blinders, Trust Tree'
+        audienceId: audienceIds.join(','), // eg. '7527562222,7527111138'
+        audienceName: audienceNames.join(','), // eg. 'Peaky Blinders, Trust Tree'
         isInCampaignHoldback,
       };
 
@@ -111,9 +122,17 @@ class Optimizely {
   };
 
   initOptimizelyIntegration(referrerOverride, sendCampaignData) {
+    const getState = () => {
+      try {
+        return window?.optimizely?.get('state');
+      } catch (e) {
+        logger.error("Couldn't successfully get state from Optimizely", e);
+      }
+    };
+
     const newActiveCampaign = (id, referrer) => {
       try {
-        const state = window?.optimizely?.get('state');
+        const state = getState();
         if (state) {
           const activeCampaigns = state.getCampaignStates({
             isActive: true,
@@ -123,17 +142,12 @@ class Optimizely {
           sendCampaignData(campaignState);
         }
       } catch (e) {
-        logger.error('Page loaded without Optimizely.');
+        logger.error("Couldn't successfully process active campaign data from Optimizely", e);
       }
     };
 
     const checkReferrer = () => {
-      let state;
-      try {
-        state = window?.optimizely?.get('state');
-      } catch (e) {
-        state = undefined;
-      }
+      const state = getState();
       if (!state) {
         return undefined;
       }
@@ -161,12 +175,7 @@ class Optimizely {
 
     const registerCurrentlyActiveCampaigns = () => {
       window.optimizely = window.optimizely || [];
-      let state;
-      try {
-        state = window?.optimizely?.get('state');
-      } catch (e) {
-        state = undefined;
-      }
+      const state = getState();
       if (state) {
         const referrer = checkReferrer();
         const activeCampaigns = state.getCampaignStates({