+++ packages/analytics-js-integrations/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.20.1](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.20.0...@rudderstack/analytics-js-integrations@3.20.1) (2026-03-02)
+
+
+### Bug Fixes
+
+* add utility to get value and fix code wherever wrongly implemented ([#2757](https://github.com/rudderlabs/rudder-sdk-js/issues/2757)) ([b60cedc](https://github.com/rudderlabs/rudder-sdk-js/commit/b60cedca80a87672218c497de139b1c20485b0b9))
+* optimizely audiences name ([#2788](https://github.com/rudderlabs/rudder-sdk-js/issues/2788)) ([8624b1a](https://github.com/rudderlabs/rudder-sdk-js/commit/8624b1a5935c3fd539aebb50f4601b8f8216f221))
+
 ## [3.20.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.19.0...@rudderstack/analytics-js-integrations@3.20.0) (2026-01-23)