+++ packages/analytics-js-plugins/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.15.3](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.15.2...@rudderstack/analytics-js-plugins@3.15.3) (2026-03-02)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-integrations` updated to version `3.20.1`
 ## [3.15.2](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.15.1...@rudderstack/analytics-js-plugins@3.15.2) (2026-02-16)
 
 ### Dependency Updates